+++ packages/polaris-viz/src/components/Annotations/utilities/checkAvailableAnnotations.ts
@@ -1,0 +1,12 @@
+import type {AnnotationLookupTable} from 'types';
+
+export function checkAvailableAnnotations(
+  annotationsLookupTable: AnnotationLookupTable,
+) {
+  const values = Object.values(annotationsLookupTable);
+
+  return {
+    hasXAxisAnnotations: values.some(({axis}) => axis === 'x'),
+    hasYAxisAnnotations: values.some(({axis}) => axis === 'y'),
+  };
+}