+++ packages/polaris-viz/src/components/StackedAreaChart/stories/StackedAreaChart.stories.tsx
@@ -185,6 +185,19 @@ const ANNOTATIONS: Annotation[] = [
     },
     axis: 'x',
   },
+  {
+    startKey: '13000',
+    label: 'Sales target',
+    axis: 'y',
+  },
+  {
+    startKey: '7500',
+    label: 'Break-even',
+    axis: 'y',
+    content: {
+      content: 'This is our break-even point. We can sell for $10 per unit.',
+    },
+  },
 ];
 
 export const Annotations: Story<StackedAreaChartProps> = Template.bind({});