+++ packages/polaris-viz/CHANGELOG.md
@@ -11,6 +11,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 - Added loading and error states to `<DonutChart/>`.
 - Added `type` prop to `<ChartSkeleton/>`.
+- Added yAxis annotations to `<LineChart />` & `<StackedAreaChart />`;
 
 ## [5.0.0] - 2022-07-07